+++ test/TransmitManager.t.sol
@@ -39,7 +39,7 @@ contract TransmitManagerTest is Setup {
     TransmitManager internal transmitManager;
 
     event SealGasLimitSet(uint256 gasLimit_);
-    event ProposeGasLimitSet(uint256 dstChainSlug_, uint256 gasLimit_);
+    event ProposeGasLimitSet(uint32 dstChainSlug_, uint256 gasLimit_);
     event TransmitManagerUpdated(address transmitManager);
     error TransmitterNotFound();
     error InsufficientTransmitFees();
