+++ test/switchboard/native/PolygonL2Switchboard.t.sol
@@ -72,7 +72,7 @@ contract PolygonL2SwitchboardTest is Setup {
 
     function deployContractsOnSingleChain(
         ChainContext storage cc_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256[] memory transmitterPrivateKeys_
     ) internal {
         // deploy socket setup
@@ -163,7 +163,7 @@ contract PolygonL2SwitchboardTest is Setup {
 
     function addPolygonL2Switchboard(
         ChainContext storage cc_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256 capacitorType_
     ) internal returns (SocketConfigContext memory scc_) {
         polygonL2Switchboard = new PolygonL2Switchboard(
@@ -190,7 +190,7 @@ contract PolygonL2SwitchboardTest is Setup {
         ChainContext storage cc_,
         address deployer_,
         address switchBoardAddress_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256 capacitorType_
     ) internal returns (SocketConfigContext memory scc_) {
         vm.startPrank(deployer_);
